Satisfy the type linting God
Okay, so this is a beeg commit.

The initial motivation was that we were getting several lint error
message like the following:

In src/components/tabs/flux.svelete:
  Object is not indexable by Type `string[]`

  <EditableValue bind:value={$cashFlows[index][item[0]]} type={item[2]} suggestions={item[3] ? item[3] : []}/>
                                                ^^^^^^^
Or, in src/store.ts:
  Type is declared but never used:

  type CashFlow {
        ^^^^^^^^

Conjugated with, in src/components/tabs/flux.svelete:
  'Unknown' Type doesn't have 'event' property

  cashFlows.subscribe((cf) => {
      cashFlowTitles[3][3] = cf.map(i => i.event).filter((k, i) => !cashFlowsBeingEdited.includes(i)).filter(unique)
                                           ^^^^^
      cashFlowTitles[4][3] = cf.map(i => i.nature).filter((k, i) => !cashFlowsBeingEdited.includes(i)).filter(unique)

And a bunch of stuff like that. All in all, I figured the root of the
issue was that we didn't give enough information to the linter to
successfully resolve the `cashFlowsTitles`'s items types (being hidden
in heterogenous arrays) and the `CashFlow` type.

So we added a type with well-typed properties for `cashFlowsTitles`,
which was renamed to the simple `columns`, exported the `CashFlow` type
from `store.ts` and added a few type hints.

We also changed the suggestions subscription, so that we don't have to
give a raw (and seemingly arbitrary) index:

-cashFlows.subscribe((cf) => {
-    cashFlowTitles[3][3] = cf.map(i => i.event).filter((k, i) => !cashFlowsBeingEdited.includes(i)).filter(unique)
-    cashFlowTitles[4][3] = cf.map(i => i.nature).filter((k, i) => !cashFlowsBeingEdited.includes(i)).filter(unique)
-})
+cashFlows.subscribe((flows: CashFlow[]) => {
+    let trackedKeys: (keyof CashFlow)[] = [ "event", "nature" ]
+    for (const key in trackedKeys) {
+        const i = columns.findIndex(c => c.id === key)
+        if (i < 0) {
+            continue
+        }
+        columns[i].suggestions = flows.map(flow => flow[key])
+            .filter((_, i) => !cashFlowsBeingEdited.includes(i))
+            .filter(unique)
+    }
+})

This approach is far less error-prone and way more robust to future
changes.

Much of the remaining diff is refoarming, adding blank lines/indents